+++ lib/rspec/matchers.rb
@@ -182,14 +182,16 @@ module RSpec
     end
 
     # Passes if actual is truthy (anything but false or nil)
-    def be_true
-      BuiltIn::BeTrue.new
+    def be_truthy
+      BuiltIn::BeTruthy.new
     end
 
     # Passes if actual is falsy (false or nil)
-    def be_false
-      BuiltIn::BeFalse.new
+    def be_falsey
+      BuiltIn::BeFalsey.new
     end
+
+    alias_method :be_falsy, :be_falsey
 
     # Passes if actual is nil
     def be_nil
@@ -197,8 +199,8 @@ module RSpec
     end
 
     # @example
-    #   expect(actual).to     be_true
-    #   expect(actual).to     be_false
+    #   expect(actual).to     be_truthy
+    #   expect(actual).to     be_falsey
     #   expect(actual).to     be_nil
     #   expect(actual).to     be_[arbitrary_predicate](*args)
     #   expect(actual).not_to be_nil